Implement start & stop commands.
